The first commit of DONG-HYUN LEE's personal game project.
This is the first version of the game project named "Slime Again".
The main file is main.py. The 'main.py' contains all of other small
parts. It is enough to show you how the all of project goes on. So
please run the 'main.py' and use others fr reference.
This file is still on the process of making. However, you can experience
several parts which are capable of running. ex. Opening parts, Modeling
parts.